Fixes for TRAFODION-3107, TRAFODION-3092, TRAFODION-3004

+++ core/sql/executor/sql_buffer.h
@@ -110,12 +110,12 @@ public:
 
   void setNextTDIOffset(ULng32 tdiOffset)
   {
-    nextTDIOffset() = (short)tdiOffset;
+    nextTDIOffset() = tdiOffset;
   }
 
   void setPrevTDIOffset(ULng32 tdiOffset)
   {
-    prevTDIOffset() = (short)tdiOffset;
+    prevTDIOffset() = tdiOffset;
   }
